31_flash_card_app/requirements.txt
pillow==12.0.0
pandas==2.3.3
